Separado widget de appbar title para pasta components
