+++ lib/app/modules/home/home_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_modular/flutter_modular.dart';
-import 'package:i9xp/app/modules/home/components/app_action_widget.dart';
+import 'package:i9xp/app/modules/home/components/appbar_home_title.dart';
 import 'package:i9xp/app/modules/home/components/category_list.dart';
 import 'package:i9xp/app/modules/home/components/category_title.dart';
 import 'package:i9xp/app/modules/home/components/latest_page_view.dart';
@@ -70,49 +70,5 @@ class _HomePageState extends ModularState<HomePage, HomeController> {
         ),
       ),
     );
-  }
-}
-
-class AppbarHomeTitle extends StatelessWidget {
-  final List<ActionModel> actions;
-  final Widget leading;
-
-  const AppbarHomeTitle({
-    Key key,
-    this.actions = const <ActionModel>[],
-    this.leading,
-  }) : super(key: key);
-
-  bool get hasActions => actions.length > 0;
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      children: [
-        leading ?? Container(),
-        if (hasActions)
-          Container(
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: _modelToWidget(),
-            ),
-          ),
-      ],
-    );
-  }
-
-  List<Widget> _modelToWidget() {
-    return actions.fold(
-      <Widget>[],
-      (List<Widget> previousValue, element) {
-        return previousValue
-          ..addAll([
-            AppAction(type: element.type, number: element.notifications),
-            SizedBox(width: 20),
-          ]);
-      },
-    ).toList()
-      ..removeLast();
   }
 }
